API

Index

C13NV.Defaults.with_defaults

C13NV.Models.make_nv_system

Reference

C13NV.Models.make_nv_system — Method

Construct the system

    H_or_L, ket, labels = make_nv_system(;
        A_zz,  # hyperfine coupling
        A_zx,  # hyperfine couplint
        B,     # magnetic field
        γ_c,   # magnetic permissibility
        δ₋,    # static detuning from |0⟩↔|-1⟩ transition
        δ₊,    # static detuning from |0⟩↔|+1⟩ transition
        ω₊,    # time-dependent frequency of |0⟩↔|+1⟩ transition
        ω₋,    # time-dependent frequency of |0⟩↔|-1⟩ transition
        Ω₋,    # time-dependent amplitude of |0⟩↔|-1⟩ transition
        Ω₊,    # time-dependent amplitude of |0⟩↔|+1⟩ transition
        Λ,     # time-dependent non-coherent excitation for the |G⟩→|E⟩ manifold
        Γ,     # spontaneous decay rate for the |E⟩→|G⟩ manifold
        Γ₀,    # spontaneous decay rate for |0E⟩→|S⟩
        Γ₊₁,   # spontaneous decay rate for |+1E⟩→|S⟩
        Γ₋₁,   # spontaneous decay rate for |-1E⟩→|S⟩
        Σ₀,    # spontaneous decay rate for |S⟩→|0G⟩
        Σ₊₁,   # spontaneous decay rate for |S⟩→|+1G⟩
        Σ₋₁,   # spontaneous decay rate for |S⟩→|-1G⟩
        use_dissipation = !isnothing(Λ)
    )

return a generator, a function ket to construct bare basis states, and a list of labels.

C13NV.Defaults.with_defaults — Method

Combine keyword arguments with default values

kwargs = with_defaults(;
    A_zz = 1.0MHz,
    A_zx = 0.3MHz,
    B = 120Gauss,
    γ_c = 1.07kHz/Gauss,
    δ₋ = 0.0,
    δ₊ = 0.0,
    Γ = (1 / (12ns)),
    Γ₀ = 0.0,
    Γ₊₁ = (1 / (24ns + 0.9ns)),
    Γ₋₁ = (1 / (24ns + 0.9ns)),
    Σ₀ = (1 / (219ns)),
    Σ₊₁ = (1 / (219ns)),
    Σ₋₁ = (1 / (219ns)),
)

accepts the keyword arguments for numerical parameter (floating point values) in C13NV.Models.make_nv_system and assigns default parameters for any value not given.