API

Index

C13NV.Defaults.with_defaults

C13NV.Models.ket

C13NV.Models.make_nv_system

Reference

C13NV.Models.ket — Method

Construct an eigenstate

Ψ = ket(label, labels; strict = true)

constructs the canonical basis state identified by label in a Hilbert space with a basis defined by labels. If label is not an element of labels, return a zero-vector if strict = false, and error if strict = true (default). The use of strict = false can be useful when working in a truncated Hilbert space.

C13NV.Models.make_nv_system — Method

Construct the system.

H_or_L, labels = make_nv_system(; kwargs...)

return a generator, and a list of labels, each label a tuple of strings.

Keyword Arguments

hyperfine_tensors: Required if A_zz, A_zx, A_zy are not given. A vector of 3×3 matrices. The number length of the vector, i.e., the number of matrices given determines the number of C-13 atoms in the model. Defaults to a single hyperfine tensor with elements A_zz, A_zx, A_zy.

A_zz, A_zx, A_zy: Required if hyperfine_tensors is not given, the strength of the hyperfine coupling along the 3 spatial axes relevant in the RWA, for a single C-13 atom.

B: Required. The magnitude of the magnetic field

ω₊ = nothing: The control $ω_{+}(t) ≡ ∂ϕ_{+}(t)/∂t$

ω₋ = nothing: The control $ω_{-}(t) ≡ ∂ϕ_{-}(t)/∂t$

Ω₊ = nothing: The control $Ω_{+}(t)$

Ω₋ = nothing: The control $Ω_{-}(t)$

Λ = nothing: The time-dependent optical drive. If given, implies the use of the full optical Hilbert space.

frame = :rwa: One of :rwa or :diag. If :diag, diagonalize the hyperfine interaction.

θ = 0.0: The azimuthal angle of the magnetic field

ϕ = 0.0: The polar angle of the magnetic field

γ_c = 1.07kHz/Gauss: The C-13 nuclear gyromagnetic ratio

δ₋ = 0.0: The detuning of the $|0⟩ ↔ |-1⟩$ transition. Defined as $δ_{-} = D - B γ_e cos(θ) - ω{-}$

δ₊ = 0.0: The detuning of the $|0⟩ ↔ |+1⟩$ transition. Defined as $δ_{+} = D - B γ_e cos(θ) - ω{+}$

Γ = 0.0: Rate for the $|E⟩ → |G⟩$ spontaneous decay

Γ₀ = 0.0: Rate for the $|E,0⟩ → |M⟩$ spontaneous decay

Γ₊₁ = 0.0: Rate for the $|E,+1⟩ → |M⟩$ spontaneous decay

Γ₋₁ = 0.0: Rate for the $|E,-1⟩ → |M⟩$ spontaneous decay

Σ₀ = 0.0: Rate for the $|M⟩ → |G,0⟩$ spontaneous decay

Σ₊₁ = 0.0: Rate for the $|M⟩ → |G,+1⟩$ spontaneous decay

Σ₋₁ = 0.0: Rate for the $|M⟩ → |G,-1⟩$ spontaneous decay

γ₊₁ = 0.0: Rate for the $|G,+1⟩ → |G,0⟩$ spontaneous decay

γ₋₁ = 0.0: Rate for the $|G,-1⟩ → |G,0⟩$ spontaneous decay

C13NV.Defaults.with_defaults — Method

Combine keyword arguments with default values

kwargs = with_defaults(;
    A_zz = 1.0MHz,
    A_zx = 0.3MHz,
    B = 120Gauss,
    γ_c = 1.07kHz/Gauss,
    δ₋ = 0.0,
    δ₊ = 0.0,
    Γ = (1 / (12ns)),
    Γ₀ = 0.0,
    Γ₊₁ = (1 / (24ns + 0.9ns)),
    Γ₋₁ = (1 / (24ns + 0.9ns)),
    Σ₀ = (1 / (219ns)),
    Σ₊₁ = (1 / (219ns)),
    Σ₋₁ = (1 / (219ns)),
)

accepts the keyword arguments for numerical parameter (floating point values) in C13NV.Models.make_nv_system and assigns default parameters for any value not given.